testing base image with config
