--- jupyterlab-dataops/resources/cloudbeaver/conf/cloudbeaver.conf
@@ -1,64 +1,0 @@
-{
-    server: {
-        serverPort: 8978,
-
-        workspaceLocation: "workspace",
-        contentRoot: "web",
-        driversLocation: "drivers",
-
-        rootURI: "/cloudbeaver",
-        serviceURI: "/api/",
-
-        productConfiguration: "conf/product.conf",
-
-        expireSessionAfterPeriod: 1800000,
-
-        develMode: false,
-
-        database: {
-            driver="h2_embedded",
-            url: "jdbc:h2:${workspace}/.data/cb.h2.dat",
-
-            createDatabase: true,
-
-            initialDataConfiguration: "conf/initial-data.conf",
-
-            pool: {
-                minIdleConnections: 4,
-                maxIdleConnections: 10,
-                maxConnections: 100,
-                validationQuery: "SELECT 1"
-            }
-        }
-
-    },
-    app: {
-        anonymousAccessEnabled: true,
-        anonymousUserRole: "admin",
-        supportsCustomConnections: true,
-
-        publicCredentialsSaveEnabled: true,
-        adminCredentialsSaveEnabled: true,
-
-        resourceQuotas: {
-            dataExportFileSizeLimit: 10000000,
-            sqlMaxRunningQueries: 3,
-            sqlResultSetRowsLimit: 100000,
-            sqlResultSetMemoryLimit: 2000000
-        }
-
-        /*
-        enabledDrivers: [
-            "postgresql:postgres-jdbc",
-            "mysql:mysql8",
-            "mysql:mariaDB",
-            "generic:sqlite_jdbc",
-            "generic:derby_server",
-            "generic:h2_embedded",
-            "jaybird:jaybird",
-        ]
-        */
-
-    }
-
-}
