--- jupyterlab-dataops/resources/cloudbeaver/conf/initial-data-sources-bigquery.conf
@@ -1,24 +1,0 @@
-{
-	"folders": {},
-	"connections": {
-		"google_bigquery_jdbc_simba-1": {
-			"provider": "bigquery",
-			"driver": "google_bigquery_jdbc_simba",
-			"name": "Google BigQuery",
-			"save-password": true,
-			"show-system-objects": true,
-			"read-only": false,
-			"configuration": {
-				"host": "https://www.googleapis.com/bigquery/v2",
-				"port": "443",
-				"url": "jdbc:bigquery://https://www.googleapis.com/bigquery/v2:443",
-				"type": "dev",
-				"properties": {
-					"OAuthPvtKeyPath": "/home/jovyan/gcp-service-account-key.json"
-				},
-				"auth-model": "native",
-				"user": "donotchange"
-			}
-		}
-	}
-}
