--- jupyterlab-dataops/resources/cloudbeaver/drivers/bigquery/pom.xml
@@ -1,53 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>drivers.bigquery</artifactId>
-    <version>1.0.0</version>
-    <parent>
-        <groupId>io.cloudbeaver</groupId>
-        <artifactId>drivers</artifactId>
-        <version>1.0.0</version>
-        <relativePath>../</relativePath>
-    </parent>
-
-    <properties>
-        <deps.output.dir>bigquery</deps.output.dir>
-    </properties>
-
-    
-    <dependencies>
-        <dependency>
-            <groupId>com.google.cloud</groupId>
-            <artifactId>google-cloud-bigquery</artifactId>
-            <version>1.127.2</version>
-        </dependency>
-        <dependency>
-            <groupId>com.google.cloud</groupId>
-            <artifactId>google-cloud-bigquerystorage</artifactId>
-            <version>1.8.5</version>
-        </dependency>
-        <dependency>
-            <groupId>com.google.cloud</groupId>
-            <artifactId>simba</artifactId>
-            <version>1.2.11.1014</version>
-            <scope>system</scope>
-            <systemPath>${project.basedir}/src/main/resources/GoogleBigQueryJDBC42.jar</systemPath>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.parquet</groupId>
-            <artifactId>parquet-avro</artifactId>
-            <version>1.10.1</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.avro</groupId>
-            <artifactId>avro</artifactId>
-            <version>1.9.0</version>
-        </dependency>
-        <dependency>
-            <groupId>joda-time</groupId>
-            <artifactId>joda-time</artifactId>
-            <version>2.9.4</version>
-        </dependency>
-    </dependencies>
-
-</project>
