--- jupyterlab-dataops/resources/cloudbeaver/plugin-default.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<plugin>
-    <!-- Resource mappings -->
-    <extension point="org.jkiss.dbeaver.resources">
-        <resource name="drivers/mysql/mysql8"/>
-        <resource name="drivers/h2"/>
-        <resource name="drivers/trino"/>
-    </extension>
-
-    <!-- Bundles  -->
-    <extension point="org.jkiss.dbeaver.product.bundles">
-        <bundle id="drivers.mysql" label="MySQL drivers"/>
-        <bundle id="drivers.h2" label="H2 drivers"/>
-        <bundle id="drivers.trino" label="Trino drivers"/>
-    </extension>
-
-    <!-- Enabled drivers -->
-    <extension point="io.cloudbeaver.driver">
-        <driver id="mysql:mysql8"/>
-        <driver id="generic:h2_embedded"/>
-        <driver id="generic:trino_jdbc"/>
-    </extension>
-
-
-</plugin>
